Added weight, height, age, and name I/O. Moved BMI calculator to function. Removed some comments. Moved Files around.
