--- Program.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using Console = System.Console;
-
-namespace BootCamp.Chapter
-{
-    class Program
-    {
-        //test comment
-        static void Main(string[] args)
-        {
-            Lesson3.Demo();
-        }
-    }
-}
