+++ Src/BootCamp.Chapter/Program .cs
@@ -1,12 +1,16 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using Console = System.Console;
 
 namespace BootCamp.Chapter
 {
     class Program
     {
+        //test comment
         static void Main(string[] args)
         {
-            Console.WriteLine("Hello World!");
+            Lesson3.Demo();
         }
     }
 }
